refactor(website): pull summoning circle setup up to name generator
